Move Schneider outlet quirk to `schneiderelectric` (#3586)
