+++ tests/test_schneiderelectric.py
@@ -2,10 +2,13 @@
 
 from unittest import mock
 
+import pytest
 from zigpy.zcl import foundation
 from zigpy.zcl.clusters.closures import WindowCovering
+from zigpy.zcl.clusters.smartenergy import Metering
 
 from tests.common import ClusterListener
+import zhaquirks.schneiderelectric.outlet
 import zhaquirks.schneiderelectric.shutters
 
 zhaquirks.setup()
@@ -135,3 +138,26 @@ async def test_1gang_shutter_1_lift_percentage_updates(zigpy_device_from_quirk):
         23,  # 100 - 77
     )
     assert len(cluster_listener.cluster_commands) == 0
+
+
+@pytest.mark.parametrize("quirk", (zhaquirks.schneiderelectric.outlet.SocketOutlet,))
+async def test_schneider_device_temp(zigpy_device_from_quirk, quirk):
+    """Test that instant demand is divided by 1000."""
+    device = zigpy_device_from_quirk(quirk)
+
+    metering_cluster = device.endpoints[6].smartenergy_metering
+    metering_listener = ClusterListener(metering_cluster)
+    instantaneous_demand_attr_id = Metering.AttributeDefs.instantaneous_demand.id
+    summation_delivered_attr_id = Metering.AttributeDefs.current_summ_delivered.id
+
+    # verify instant demand is divided by 1000
+    metering_cluster.update_attribute(instantaneous_demand_attr_id, 25000)
+    assert len(metering_listener.attribute_updates) == 1
+    assert metering_listener.attribute_updates[0][0] == instantaneous_demand_attr_id
+    assert metering_listener.attribute_updates[0][1] == 25  # divided by 1000
+
+    # verify other attributes are not modified
+    metering_cluster.update_attribute(summation_delivered_attr_id, 25)
+    assert len(metering_listener.attribute_updates) == 2
+    assert metering_listener.attribute_updates[1][0] == summation_delivered_attr_id
+    assert metering_listener.attribute_updates[1][1] == 25  # not modified
