--- zhaquirks/schneider/__init__.py
@@ -1,3 +1,0 @@
-"""Quirks for Schneider devices."""
-
-SCHNEIDER = "Schneider Electric"
